create separate class for each API, update Allure report
